chore: change seeding method
