+++ package.json
@@ -9,7 +9,10 @@
     "db:migrate": "drizzle-kit migrate",
     "db:push": "drizzle-kit push",
     "db:studio": "drizzle-kit studio",
-    "db:seed": "dotenv npx tsx ./src/server/db/seed.ts",
+    "db:seed:educations": "dotenv npx tsx ./src/server/db/seed/educations.seed.ts",
+    "db:seed:license": "dotenv npx tsx ./src/server/db/seed/license.seed.ts",
+    "db:seed:users": "dotenv npx tsx ./src/server/db/seed/user.seed.ts",
+    "db:seed:all": "dotenv npx tsx ./src/server/db/seed/index.ts",
     "db:schema:generate": "node --experimental-specifier-resolution=node --loader ts-node/esm src/server/db/dbml-generator/dbml.ts && dbml-renderer -i schema.dbml -o erd.svg",
     "dev": "next dev",
     "lint": "next lint",
